```diff + Moved some assistant-like options to their own menu. Access from Top Bar -> More + ```
