+++ src/ClassicUO.Client/Game/UI/Gumps/ModernOptionsGump.cs
@@ -3812,75 +3812,6 @@ namespace ClassicUO.Game.UI.Gumps
 
             #endregion
 
-            #region Autoloot
-
-            page = ((int)PAGE.TUOOptions + 1012);
-            content.AddToLeft(SubCategoryButton(lang.GetTazUO.AutoLoot, page, content.LeftWidth));
-            content.ResetRightSide();
-
-            content.AddToRight(new HttpClickableLink("Autoloot Wiki", "https://github.com/bittiez/TazUO/wiki/TazUO.Simple-Auto-Loot", ThemeSettings.TEXT_FONT_COLOR), true, page);
-            content.BlankLine();
-
-            content.AddToRight(new CheckboxWithLabel(lang.GetTazUO.AutoLootEnable, 0, profile.EnableAutoLoot, b => profile.EnableAutoLoot = b), true, page);
-            content.BlankLine();
-
-            content.AddToRight(new CheckboxWithLabel(lang.GetTazUO.ScavengerEnable, 0, profile.EnableScavenger, b => profile.EnableScavenger = b), true, page);
-            content.BlankLine();
-
-            content.AddToRight
-                (new CheckboxWithLabel(lang.GetTazUO.AutoLootProgessBarEnable, 0, profile.EnableAutoLootProgressBar, b => profile.EnableAutoLootProgressBar = b), true, page);
-
-            content.BlankLine();
-
-            content.AddToRight(new CheckboxWithLabel(lang.GetTazUO.AutoLootHumanCorpses, 0, profile.AutoLootHumanCorpses, b => profile.AutoLootHumanCorpses = b), true, page);
-            content.BlankLine();
-
-            content.AddToRight(c = new AutoLootConfigs(content.RightWidth - ThemeSettings.SCROLL_BAR_WIDTH - 10), true, page);
-
-            #endregion
-
-            #region Auto sell
-
-            page = ((int)PAGE.TUOOptions + 1013);
-            content.AddToLeft(SubCategoryButton(lang.GetTazUO.AutoSellMenu, page, content.LeftWidth));
-            content.ResetRightSide();
-
-            content.AddToRight(new HttpClickableLink("Auto Sell Wiki", "https://github.com/bittiez/TazUO/wiki/TazUO.Auto-Sell-Agent", ThemeSettings.TEXT_FONT_COLOR), true, page);
-            content.BlankLine();
-
-            content.AddToRight(new CheckboxWithLabel(lang.GetTazUO.AutoSellEnable, 0, profile.SellAgentEnabled, b => profile.SellAgentEnabled = b), true, page);
-            content.BlankLine();
-
-            content.AddToRight(new SellAgentConfigs(content.RightWidth - ThemeSettings.SCROLL_BAR_WIDTH - 10), true, page);
-
-            #endregion
-
-            #region Auto buy
-
-            page = ((int)PAGE.TUOOptions + 1014);
-            content.AddToLeft(SubCategoryButton(lang.GetTazUO.AutoBuyMenu, page, content.LeftWidth));
-            content.ResetRightSide();
-
-            content.AddToRight(new HttpClickableLink("Auto Buy Wiki", "https://github.com/bittiez/TazUO/wiki/TazUO.Auto-Buy-Agent", ThemeSettings.TEXT_FONT_COLOR), true, page);
-            content.BlankLine();
-
-            content.AddToRight(new CheckboxWithLabel(lang.GetTazUO.AutoBuyEnable, 0, profile.BuyAgentEnabled, b => profile.BuyAgentEnabled = b), true, page);
-            content.BlankLine();
-
-            content.AddToRight(new BuyAgentConfigs(content.RightWidth - ThemeSettings.SCROLL_BAR_WIDTH - 10), true, page);
-
-            #endregion
-
-            #region Graphic Filter
-
-            page = ((int)PAGE.TUOOptions + 1015);
-            content.AddToLeft(SubCategoryButton(lang.GetTazUO.GraphicChangeFilter, page, content.LeftWidth));
-            content.ResetRightSide();
-
-            content.AddToRight(new GraphicFilterConfigs(content.RightWidth - ThemeSettings.SCROLL_BAR_WIDTH - 10), true, page);
-
-            #endregion
-
             #region Hotkeys
 
             page = ((int)PAGE.TUOOptions + 1016);
@@ -4602,796 +4533,6 @@ namespace ClassicUO.Game.UI.Gumps
                 dataBox.ReArrangeChildren(2);
                 dataBox.ForceSizeUpdate();
                 ForceSizeUpdate();
-            }
-        }
-
-        private class GraphicFilterConfigs : Control
-        {
-            private DataBox _dataBox;
-
-            public GraphicFilterConfigs(int width)
-            {
-                AcceptMouseInput = true;
-                CanMove = true;
-                Width = width;
-
-                Add(_dataBox = new DataBox(0, 0, width, 0));
-
-                ModernButton b;
-                _dataBox.Add(b = new ModernButton(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Add blank entry", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    var newConfig = GraphicsReplacement.NewFilter(0, 0);
-
-                    if (newConfig != null)
-                    {
-                        _dataBox.Insert(3, GenConfigEntry(newConfig, width));
-                        RearrangeDataBox();
-                    }
-                };
-
-                _dataBox.Add(b = new ModernButton(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Target entity", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    TargetHelper.TargetObject
-                    ((e) =>
-                        {
-                            if (e == null)
-                                return;
-
-                            // if (e == null || !SerialHelper.IsMobile(e)) return;
-                            var sc = GraphicsReplacement.NewFilter(e.Graphic, e.Graphic, e.Hue);
-
-                            if (sc != null && _dataBox != null)
-                            {
-                                _dataBox.Insert(3, GenConfigEntry(sc, width));
-                                RearrangeDataBox();
-                            }
-                        }
-                    );
-                };
-
-                Area titles = new Area(false);
-
-                Control c;
-                titles.Add(TextBox.GetOne("Graphic", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default()));
-                titles.Add(c = TextBox.GetOne("New Graphic", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default()));
-                c.X = ((width - 90 - 5) / 3) + 5;
-                titles.Add(c = TextBox.GetOne("New Hue", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default()));
-                c.X = (((width - 90 - 5) / 3) * 2) + 10;
-                titles.ForceSizeUpdate();
-                _dataBox.Add(titles);
-
-                foreach (var item in GraphicsReplacement.GraphicFilters)
-                {
-                    _dataBox.Add(GenConfigEntry(item.Value, width));
-                }
-
-                RearrangeDataBox();
-            }
-
-            private Control GenConfigEntry(GraphicChangeFilter filter, int width)
-            {
-                int ewidth = (width - 90) / 3;
-
-                Area area = new Area()
-                {
-                    Width = width,
-                    Height = 50
-                };
-
-                int x = 0;
-
-                InputField graphicInput = new InputField
-                (
-                    ewidth, 50, 100, -1, filter.OriginalGraphic.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox graphicInput = (InputField.StbTextBox)s;
-
-                        if (graphicInput.Text.StartsWith("0x") && ushort.TryParse(graphicInput.Text.Substring(2), NumberStyles.AllowHexSpecifier, null, out var ngh))
-                        {
-                            filter.OriginalGraphic = ngh;
-                            GraphicsReplacement.ResetLists();
-                        }
-                        else if (ushort.TryParse(graphicInput.Text, out var ng))
-                        {
-                            filter.OriginalGraphic = ng;
-                            GraphicsReplacement.ResetLists();
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                graphicInput.SetTooltip("Original Graphic");
-                area.Add(graphicInput);
-                x += graphicInput.Width + 5;
-
-                InputField newgraphicInput = new InputField
-                (
-                    ewidth, 50, 100, -1, filter.ReplacementGraphic.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox graphicInput = (InputField.StbTextBox)s;
-
-                        if (graphicInput.Text.StartsWith("0x") && ushort.TryParse(graphicInput.Text.Substring(2), NumberStyles.AllowHexSpecifier, null, out var ngh))
-                        {
-                            filter.ReplacementGraphic = ngh;
-                        }
-                        else if (ushort.TryParse(graphicInput.Text, out var ng))
-                        {
-                            filter.ReplacementGraphic = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                newgraphicInput.SetTooltip("Replacement Graphic");
-                area.Add(newgraphicInput);
-                x += newgraphicInput.Width + 5;
-
-                InputField hueInput = new InputField
-                (
-                    ewidth, 50, 100, -1, filter.NewHue == ushort.MaxValue ? "-1" : filter.NewHue.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox hueInput = (InputField.StbTextBox)s;
-
-                        if (hueInput.Text == "-1")
-                        {
-                            filter.NewHue = ushort.MaxValue;
-                        }
-                        else if (ushort.TryParse(hueInput.Text, out var ng))
-                        {
-                            filter.NewHue = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                hueInput.SetTooltip("Hue (-1 to leave original)");
-                area.Add(hueInput);
-                x += hueInput.Width + 5;
-
-                NiceButton delete;
-
-                area.Add
-                (
-                    delete = new NiceButton(x, 0, area.Width - x, 49, ButtonAction.Activate, "X")
-                    {
-                        IsSelectable = false,
-                        DisplayBorder = true
-                    }
-                );
-
-                delete.SetTooltip("Delete this entry");
-
-                delete.MouseUp += (s, e) =>
-                {
-                    if (e.Button == Input.MouseButtonType.Left)
-                    {
-                        GraphicsReplacement.DeleteFilter(filter.OriginalGraphic);
-                        area.Dispose();
-                        RearrangeDataBox();
-                    }
-                };
-
-                return area;
-            }
-
-            private void RearrangeDataBox()
-            {
-                _dataBox.ReArrangeChildren();
-                _dataBox.ForceSizeUpdate();
-                Height = _dataBox.Height;
-            }
-        }
-
-        private class BuyAgentConfigs : Control
-        {
-            private DataBox _dataBox;
-
-            public BuyAgentConfigs(int width)
-            {
-                AcceptMouseInput = true;
-                CanMove = true;
-                Width = width;
-
-                Add(_dataBox = new DataBox(0, 0, width, 0));
-
-                ModernButton b;
-                _dataBox.Add(b = new ModernButton(0, 0, 100, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Add entry", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    _dataBox.Insert(3, GenConfigEntry(BuySellAgent.Instance.NewBuyConfig(), width));
-                    RearrangeDataBox();
-                };
-
-                _dataBox.Add(b = new ModernButton(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Target item", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    TargetHelper.TargetObject
-                    ((e) =>
-                        {
-                            if (e == null)
-                                return;
-
-                            var sc = BuySellAgent.Instance.NewBuyConfig();
-                            sc.Graphic = e.Graphic;
-                            sc.Hue = e.Hue;
-                            _dataBox.Insert(3, GenConfigEntry(sc, width));
-                            RearrangeDataBox();
-                        }
-                    );
-                };
-
-                Area titles = new Area(false);
-
-                TextBox tempTextBox1 = TextBox.GetOne
-                    ("Graphic", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default(null));
-
-                tempTextBox1.X = 50;
-                titles.Add(tempTextBox1);
-
-                tempTextBox1 = TextBox.GetOne("Hue", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default(null));
-                tempTextBox1.X = ((width - 90 - 60) / 3) + 55;
-                titles.Add(tempTextBox1);
-
-                tempTextBox1 = TextBox.GetOne("Max Amount", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default(null));
-                tempTextBox1.X = (((width - 90 - 60) / 3) * 2) + 60;
-                titles.Add(tempTextBox1);
-
-                titles.ForceSizeUpdate();
-                _dataBox.Add(titles);
-
-                if (BuySellAgent.Instance.BuyConfigs != null)
-                    foreach (var item in BuySellAgent.Instance.BuyConfigs)
-                    {
-                        _dataBox.Add(GenConfigEntry(item, width));
-                    }
-
-                RearrangeDataBox();
-            }
-
-            private Control GenConfigEntry(BuySellItemConfig itemConfig, int width)
-            {
-                int ewidth = (width - 90 - 60) / 3;
-
-                Area area = new Area()
-                {
-                    Width = width,
-                    Height = 50
-                };
-
-                int x = 0;
-
-                if (itemConfig.Graphic > 0)
-                {
-                    ResizableStaticPic rsp;
-
-                    area.Add
-                    (
-                        rsp = new ResizableStaticPic(itemConfig.Graphic, 50, 50)
-                        {
-                            Hue = (ushort)(itemConfig.Hue == ushort.MaxValue ? 0 : itemConfig.Hue)
-                        }
-                    );
-                }
-
-                x += 50;
-
-                InputField graphicInput = new InputField
-                (
-                    ewidth, 50, 100, -1, itemConfig.Graphic.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox graphicInput = (InputField.StbTextBox)s;
-
-                        if (graphicInput.Text.StartsWith("0x") && ushort.TryParse(graphicInput.Text.Substring(2), NumberStyles.AllowHexSpecifier, null, out var ngh))
-                        {
-                            itemConfig.Graphic = ngh;
-                        }
-                        else if (ushort.TryParse(graphicInput.Text, out var ng))
-                        {
-                            itemConfig.Graphic = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                graphicInput.SetTooltip("Graphic");
-                area.Add(graphicInput);
-                x += graphicInput.Width + 5;
-
-
-                InputField hueInput = new InputField
-                (
-                    ewidth, 50, 100, -1, itemConfig.Hue == ushort.MaxValue ? "-1" : itemConfig.Hue.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox hueInput = (InputField.StbTextBox)s;
-
-                        if (hueInput.Text == "-1")
-                        {
-                            itemConfig.Hue = ushort.MaxValue;
-                        }
-                        else if (ushort.TryParse(hueInput.Text, out var ng))
-                        {
-                            itemConfig.Hue = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                hueInput.SetTooltip("Hue (-1 to match any)");
-                area.Add(hueInput);
-                x += hueInput.Width + 5;
-
-                InputField maxInput = new InputField
-                (
-                    ewidth, 50, 100, -1, itemConfig.MaxAmount.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox maxInput = (InputField.StbTextBox)s;
-
-                        if (ushort.TryParse(maxInput.Text, out var ng))
-                        {
-                            itemConfig.MaxAmount = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                maxInput.SetTooltip("Max Amount");
-                area.Add(maxInput);
-                x += maxInput.Width + 5;
-
-                CheckboxWithLabel enabled = new CheckboxWithLabel(isChecked: itemConfig.Enabled, valueChanged: (e) => { itemConfig.Enabled = e; })
-                {
-                    X = x
-                };
-
-                enabled.Y = (area.Height - enabled.Height) >> 1;
-                enabled.SetTooltip("Enable this entry?");
-                area.Add(enabled);
-                x += enabled.Width;
-
-                NiceButton delete;
-
-                area.Add
-                (
-                    delete = new NiceButton(x, 0, area.Width - x, 49, ButtonAction.Activate, "X")
-                    {
-                        IsSelectable = false,
-                        DisplayBorder = true
-                    }
-                );
-
-                delete.SetTooltip("Delete this entry");
-
-                delete.MouseUp += (s, e) =>
-                {
-                    if (e.Button == Input.MouseButtonType.Left)
-                    {
-                        BuySellAgent.Instance?.DeleteConfig(itemConfig);
-                        area.Dispose();
-                        RearrangeDataBox();
-                    }
-                };
-
-                return area;
-            }
-
-            private void RearrangeDataBox()
-            {
-                _dataBox.ReArrangeChildren();
-                _dataBox.ForceSizeUpdate();
-                Height = _dataBox.Height;
-            }
-        }
-
-        private class SellAgentConfigs : Control
-        {
-            private DataBox _dataBox;
-
-            public SellAgentConfigs(int width)
-            {
-                AcceptMouseInput = true;
-                CanMove = true;
-                Width = width;
-
-                Add(_dataBox = new DataBox(0, 0, width, 0));
-
-                ModernButton b;
-                _dataBox.Add(b = new ModernButton(0, 0, 100, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Add entry", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    _dataBox.Insert(3, GenConfigEntry(BuySellAgent.Instance.NewSellConfig(), width));
-                    RearrangeDataBox();
-                };
-
-                _dataBox.Add(b = new ModernButton(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Target item", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    TargetHelper.TargetObject
-                    ((e) =>
-                        {
-                            if (e == null)
-                                return;
-
-                            var sc = BuySellAgent.Instance.NewSellConfig();
-                            sc.Graphic = e.Graphic;
-                            sc.Hue = e.Hue;
-                            _dataBox.Insert(3, GenConfigEntry(sc, width));
-                            RearrangeDataBox();
-                        }
-                    );
-                };
-
-                Area titles = new Area(false);
-
-                TextBox tempTextBox1 = TextBox.GetOne("Graphic", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default());
-                tempTextBox1.X = 50;
-                titles.Add(tempTextBox1);
-
-                tempTextBox1 = TextBox.GetOne("Hue", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default());
-                tempTextBox1.X = ((width - 90 - 60) / 3) + 55;
-                titles.Add(tempTextBox1);
-
-                tempTextBox1 = TextBox.GetOne("Max Amount", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default());
-                tempTextBox1.X = (((width - 90 - 60) / 3) * 2) + 60;
-                titles.Add(tempTextBox1);
-
-                titles.ForceSizeUpdate();
-                _dataBox.Add(titles);
-
-                if (BuySellAgent.Instance.SellConfigs != null)
-                    foreach (var item in BuySellAgent.Instance.SellConfigs)
-                    {
-                        _dataBox.Add(GenConfigEntry(item, width));
-                    }
-
-                RearrangeDataBox();
-            }
-
-            private Control GenConfigEntry(BuySellItemConfig itemConfig, int width)
-            {
-                int ewidth = (width - 90 - 60) / 3;
-
-                Area area = new Area()
-                {
-                    Width = width,
-                    Height = 50
-                };
-
-                int x = 0;
-
-                if (itemConfig.Graphic > 0)
-                {
-                    ResizableStaticPic rsp;
-
-                    area.Add
-                    (
-                        rsp = new ResizableStaticPic(itemConfig.Graphic, 50, 50)
-                        {
-                            Hue = (ushort)(itemConfig.Hue == ushort.MaxValue ? 0 : itemConfig.Hue)
-                        }
-                    );
-                }
-
-                x += 50;
-
-                InputField graphicInput = new InputField
-                (
-                    ewidth, 50, 100, -1, itemConfig.Graphic.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox graphicInput = (InputField.StbTextBox)s;
-
-                        if (graphicInput.Text.StartsWith("0x") && ushort.TryParse(graphicInput.Text.Substring(2), NumberStyles.AllowHexSpecifier, null, out var ngh))
-                        {
-                            itemConfig.Graphic = ngh;
-                        }
-                        else if (ushort.TryParse(graphicInput.Text, out var ng))
-                        {
-                            itemConfig.Graphic = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                graphicInput.SetTooltip("Graphic");
-                area.Add(graphicInput);
-                x += graphicInput.Width + 5;
-
-
-                InputField hueInput = new InputField
-                (
-                    ewidth, 50, 100, -1, itemConfig.Hue == ushort.MaxValue ? "-1" : itemConfig.Hue.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox hueInput = (InputField.StbTextBox)s;
-
-                        if (hueInput.Text == "-1")
-                        {
-                            itemConfig.Hue = ushort.MaxValue;
-                        }
-                        else if (ushort.TryParse(hueInput.Text, out var ng))
-                        {
-                            itemConfig.Hue = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                hueInput.SetTooltip("Hue (-1 to match any)");
-                area.Add(hueInput);
-                x += hueInput.Width + 5;
-
-                InputField maxInput = new InputField
-                (
-                    ewidth, 50, 100, -1, itemConfig.MaxAmount.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox maxInput = (InputField.StbTextBox)s;
-
-                        if (ushort.TryParse(maxInput.Text, out var ng))
-                        {
-                            itemConfig.MaxAmount = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                maxInput.SetTooltip("Max Amount");
-                area.Add(maxInput);
-                x += maxInput.Width + 5;
-
-                CheckboxWithLabel enabled = new CheckboxWithLabel(isChecked: itemConfig.Enabled, valueChanged: (e) => { itemConfig.Enabled = e; })
-                {
-                    X = x
-                };
-
-                enabled.Y = (area.Height - enabled.Height) >> 1;
-                enabled.SetTooltip("Enable this entry?");
-                area.Add(enabled);
-                x += enabled.Width;
-
-                NiceButton delete;
-
-                area.Add
-                (
-                    delete = new NiceButton(x, 0, area.Width - x, 49, ButtonAction.Activate, "X")
-                    {
-                        IsSelectable = false,
-                        DisplayBorder = true
-                    }
-                );
-
-                delete.SetTooltip("Delete this entry");
-
-                delete.MouseUp += (s, e) =>
-                {
-                    if (e.Button == Input.MouseButtonType.Left)
-                    {
-                        BuySellAgent.Instance?.DeleteConfig(itemConfig);
-                        area.Dispose();
-                        RearrangeDataBox();
-                    }
-                };
-
-                return area;
-            }
-
-            private void RearrangeDataBox()
-            {
-                _dataBox.ReArrangeChildren();
-                _dataBox.ForceSizeUpdate();
-                Height = _dataBox.Height;
-            }
-        }
-
-        private class AutoLootConfigs : Control
-        {
-            private DataBox _dataBox;
-
-            public AutoLootConfigs(int width)
-            {
-                AcceptMouseInput = true;
-                CanMove = true;
-                Width = width;
-
-                Add(_dataBox = new DataBox(0, 0, width, 0));
-
-                ModernButton b;
-                _dataBox.Add(b = new ModernButton(0, 0, 100, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Add entry", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    var nl = AutoLootManager.Instance.AddAutoLootEntry();
-                    _dataBox.Insert(2, GenConfigEntry(nl, width));
-                    RearrangeDataBox();
-                };
-
-                _dataBox.Add(b = new ModernButton(0, 0, 200, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Target item to add", ThemeSettings.BUTTON_FONT_COLOR));
-
-                b.MouseUp += (s, e) =>
-                {
-                    TargetHelper.TargetObject
-                    ((o) =>
-                        {
-                            if (o != null)
-                            {
-                                var nl = AutoLootManager.Instance.AddAutoLootEntry(o.Graphic, o.Hue, o.Name);
-
-                                if (_dataBox != null)
-                                {
-                                    _dataBox.Insert(2, GenConfigEntry(nl, width));
-                                    RearrangeDataBox();
-                                }
-                            }
-                        }
-                    );
-                };
-
-                Area titles = new Area(false);
-                TextBox tempTextBox1 = TextBox.GetOne("Graphic", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default());
-                tempTextBox1.X = 55;
-                titles.Add(tempTextBox1);
-
-                tempTextBox1 = TextBox.GetOne("Hue", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default());
-                tempTextBox1.X = ((width - 90 - 50) >> 1) + 60;
-                titles.Add(tempTextBox1);
-
-                titles.ForceSizeUpdate();
-                _dataBox.Add(titles);
-
-                for (int i = 0; i < AutoLootManager.Instance.AutoLootList.Count; i++)
-                {
-                    AutoLootConfigEntry autoLootItem = AutoLootManager.Instance.AutoLootList[i];
-                    _dataBox.Add(GenConfigEntry(autoLootItem, width));
-                }
-
-                RearrangeDataBox();
-            }
-
-            private Control GenConfigEntry(AutoLootConfigEntry autoLootItem, int width)
-            {
-                int ewidth = (width - 90 - 60) >> 1;
-
-                Area area = new Area()
-                {
-                    Width = width,
-                    Height = 107
-                };
-
-                int x = 0;
-
-                if (autoLootItem.Graphic > 0)
-                {
-                    ResizableStaticPic rsp;
-
-                    area.Add
-                    (
-                        rsp = new ResizableStaticPic((uint)autoLootItem.Graphic, 50, 50)
-                        {
-                            Hue = (ushort)(autoLootItem.Hue == ushort.MaxValue ? 0 : autoLootItem.Hue)
-                        }
-                    );
-
-                    rsp.SetTooltip(autoLootItem.Name);
-                }
-
-                x += 50;
-
-                InputField graphicInput = new InputField
-                (
-                    ewidth, 50, 100, -1, autoLootItem.Graphic.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox graphicInput = (InputField.StbTextBox)s;
-
-                        if (graphicInput.Text.StartsWith("0x") && short.TryParse(graphicInput.Text.Substring(2), NumberStyles.AllowHexSpecifier, null, out var ngh))
-                        {
-                            autoLootItem.Graphic = ngh;
-                        }
-                        else if (int.TryParse(graphicInput.Text, out var ng))
-                        {
-                            autoLootItem.Graphic = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                graphicInput.SetTooltip("Graphic");
-                area.Add(graphicInput);
-                x += graphicInput.Width + 5;
-
-
-                InputField hueInput = new InputField
-                (
-                    ewidth, 50, 100, -1, autoLootItem.Hue == ushort.MaxValue ? "-1" : autoLootItem.Hue.ToString(), false, (s, e) =>
-                    {
-                        InputField.StbTextBox hueInput = (InputField.StbTextBox)s;
-
-                        if (hueInput.Text == "-1")
-                        {
-                            autoLootItem.Hue = ushort.MaxValue;
-                        }
-                        else if (ushort.TryParse(hueInput.Text, out var ng))
-                        {
-                            autoLootItem.Hue = ng;
-                        }
-                    }
-                )
-                {
-                    X = x
-                };
-
-                hueInput.SetTooltip("Hue (-1 to match any)");
-                area.Add(hueInput);
-                x += hueInput.Width + 5;
-
-                NiceButton delete;
-
-                area.Add
-                (
-                    delete = new NiceButton(x, 0, 90, 49, ButtonAction.Activate, "Delete")
-                    {
-                        IsSelectable = false,
-                        DisplayBorder = true
-                    }
-                );
-
-                delete.MouseUp += (s, e) =>
-                {
-                    if (e.Button == Input.MouseButtonType.Left)
-                    {
-                        AutoLootManager.Instance.TryRemoveAutoLootEntry(autoLootItem.UID);
-                        area.Dispose();
-                        RearrangeDataBox();
-                    }
-                };
-
-                InputField regxInput = new InputField
-                (
-                    width, 50, width, -1, autoLootItem.RegexSearch, false, (s, e) =>
-                    {
-                        InputField.StbTextBox regxInput = (InputField.StbTextBox)s;
-                        autoLootItem.RegexSearch = string.IsNullOrEmpty(regxInput.Text) ? string.Empty : regxInput.Text;
-                    }
-                )
-                {
-                    Y = 52
-                };
-
-                regxInput.SetTooltip("Regex to match items against");
-                area.Add(regxInput);
-
-                return area;
-            }
-
-            private void RearrangeDataBox()
-            {
-                _dataBox.ReArrangeChildren();
-                _dataBox.ForceSizeUpdate();
-                Height = _dataBox.Height;
             }
         }
